requirements.txt
streamlit>=1.24.0
openai>=1.0.0
Pillow>=9.0.0
pdf2image>=1.16.0
faiss-cpu>=1.7.0
pytesseract>=0.3.0
langdetect>=1.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.0
numpy>=1.21.0
typing>=3.7.4